requirements.txt
fastapi==0.112.0
uvicorn==0.15.0
python-dotenv==1.0.1
#langchain-core==0.3.0
langchain_aws==0.2.0
boto3==1.34.162
langchain-google-vertexai==2.0.0